visao-computacional-em-producao/requirements.txt
fastapi
mlflow
opencv-python
pillow
torch
torchvision
ultralytics
uvicorn
